Fix warning

+++ src/device.cpp
@@ -434,7 +434,7 @@ void Device::ExecuteTasks(Submission& tasks)
                 {
                     tasks[i]->Complete(CL_OUT_OF_RESOURCES, Lock);
                 }
-                for (int j = i + 1; j < tasks.size(); ++j)
+                for (size_t j = i + 1; j < tasks.size(); ++j)
                 {
                     auto& task = tasks[j];
                     task->Complete(CL_EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST, Lock);
